@@ -17,9 +17,8 @@
     "build": "ng build picker",
     "test": "ng test picker",
     "lint": "ng lint",
-    "e2e": "ng e2e",
     "build_lib": "ng build picker && cp README.md dist/picker",
-    "build_css": "mkdir -p dist/picker/assets/style && node-sass --output-style=compressed projects/picker/src/sass/picker.scss > dist/picker/assets/style/picker.min.css",
+    "build_css": "mkdir -p dist/picker/assets/style && sass --style=compressed projects/picker/src/sass/picker.scss:dist/picker/assets/style/picker.min.css",
     "publish": "cd dist/picker && npm publish",
     "package": "npm run build_lib && npm run build_css"
   },
@@ -28,45 +27,44 @@
     "url": "https://github.com/DanielYKPan/date-time-picker.git"
   },
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "moment-timezone": "^0.5.27",
-    "rxjs": "~6.5.3",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "~13.2.2",
+    "@angular/cdk": "~13.2.2",
+    "@angular/common": "~13.2.2",
+    "@angular/compiler": "~13.2.2",
+    "@angular/core": "~13.2.2",
+    "@angular/forms": "~13.2.2",
+    "@angular/platform-browser": "~13.2.2",
+    "@angular/platform-browser-dynamic": "~13.2.2",
+    "@angular/router": "~13.2.2",
+    "rxjs": "~7.5.4",
+    "tslib": "^2.3.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.20",
-    "@angular-devkit/build-ng-packagr": "~0.803.20",
-    "@angular/cdk": "~8.2.3",
-    "@angular/cli": "~8.3.20",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/jasmine": "^3.3.16",
-    "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^11.13.23",
-    "codelyzer": "^5.2.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.6",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "ng-packagr": "^5.7.1",
-    "node-sass": "^4.13.0",
-    "prettier": "1.19.1",
-    "protractor": "~5.4.2",
-    "ts-node": "^8.1.1",
-    "tsickle": "0.37.0",
-    "tslib": "^1.10.0",
-    "tslint": "~5.16.0",
-    "tslint-config-prettier": "^1.18.0",
-    "typescript": "~3.5.3"
+    "@angular-devkit/build-angular": "~13.2.3",
+    "@angular-eslint/builder": "^13.1.0",
+    "@angular-eslint/eslint-plugin": "^13.1.0",
+    "@angular-eslint/eslint-plugin-template": "^13.1.0",
+    "@angular-eslint/schematics": "^13.1.0",
+    "@angular-eslint/template-parser": "^13.1.0",
+    "@angular/cli": "~13.2.3",
+    "@angular/compiler-cli": "~13.2.2",
+    "@types/jasmine": "~3.10.3",
+    "@types/node": "^16.11.24",
+    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/parser": "^5.12.0",
+    "eslint": "^8.9.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "jasmine-core": "~4.0.0",
+    "karma": "~6.3.16",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.1.1",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "ng-packagr": "^13.2.1",
+    "prettier": "~2.5.1",
+    "sass": "^1.49.7",
+    "typescript": "~4.5.5"
   }
 }
