@@ -1,6 +1,6 @@
 {
   "name": "ng-date-and-time-picker",
-  "version": "8.1.2",
+  "version": "13.1.0",
   "description": "Angular Date Time Picker",
   "keywords": [
     "Angular",
@@ -16,13 +16,12 @@
     "url": "https://github.com/DanielYKPan/date-time-picker.git"
   },
   "peerDependencies": {
-    "@angular/common": "^8.0.0",
-    "@angular/core": "^8.0.0",
-    "@angular/cdk": "^8.0.0",
-    "rxjs": "^6.5.3",
-    "moment": "^2.24.0"
+    "@angular/common": "^13.2.2",
+    "@angular/core": "^13.2.2",
+    "@angular/cdk": "^13.2.2",
+    "rxjs": "^7.5.4"
   },
-  "optionalDependencies": {
-    "moment": "^2.24.0"
+  "dependencies": {
+    "tslib": "^2.3.1"
   }
 }
