@@ -1,118 +1,128 @@
-/**
- * timer-box.component
- */
-
 import {
-    ChangeDetectionStrategy,
-    Component,
-    EventEmitter,
-    Input,
-    OnDestroy,
-    OnInit,
-    Output
+  ChangeDetectionStrategy,
+  Component,
+  EventEmitter,
+  Input,
+  OnDestroy,
+  OnInit,
+  Output
 } from '@angular/core';
 import { coerceNumberProperty } from '@angular/cdk/coercion';
 import { Subject, Subscription } from 'rxjs';
-import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { debounceTime } from 'rxjs/operators';
 
 @Component({
-    exportAs: 'owlDateTimeTimerBox',
-    selector: 'owl-date-time-timer-box',
-    templateUrl: './timer-box.component.html',
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    host: {
-        '[class.owl-dt-timer-box]': 'owlDTTimerBoxClass'
-    }
+  exportAs: 'owlDateTimeTimerBox',
+  selector: 'owl-date-time-timer-box',
+  templateUrl: './timer-box.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  // eslint-disable-next-line @angular-eslint/no-host-metadata-property
+  host: {
+    '[class.owl-dt-timer-box]': 'owlDTTimerBoxClass'
+  }
 })
 export class OwlTimerBoxComponent implements OnInit, OnDestroy {
-    @Input() showDivider = false;
+  @Input() showDivider = false;
 
-    @Input() upBtnAriaLabel: string;
+  @Input() upBtnAriaLabel: string;
 
-    @Input() upBtnDisabled: boolean;
+  @Input() upBtnDisabled: boolean;
 
-    @Input() downBtnAriaLabel: string;
+  @Input() downBtnAriaLabel: string;
 
-    @Input() downBtnDisabled: boolean;
+  @Input() downBtnDisabled: boolean;
 
-    /**
-     * Value would be displayed in the box
-     * If it is null, the box would display [value]
-     */
-    @Input() boxValue: number;
+  /**
+   * Value would be displayed in the box
+   * If it is null, the box would display [value]
+   */
+  @Input() boxValue: number;
 
-    @Input() value: number;
+  @Input() value: number;
 
-    @Input() min: number;
+  @Input() min: number;
 
-    @Input() max: number;
+  @Input() max: number;
 
-    @Input() step = 1;
+  @Input() step = 1;
 
-    @Input() inputLabel: string;
+  @Input() inputLabel: string;
 
-    @Output() valueChange = new EventEmitter<number>();
+  @Output() valueChange = new EventEmitter<number>();
 
-    @Output() inputChange = new EventEmitter<number>();
+  @Output() inputChange = new EventEmitter<number>();
 
-    private inputStream = new Subject<string>();
+  private inputStream = new Subject<string>();
 
-    private inputStreamSub = Subscription.EMPTY;
+  private inputStreamSub = Subscription.EMPTY;
 
-    constructor() {}
+  constructor() {}
 
-    get displayValue(): number {
-        return this.boxValue || this.value;
+  get displayValue(): number {
+    return this.boxValue || this.value;
+  }
+
+  get owlDTTimerBoxClass(): boolean {
+    return true;
+  }
+
+  public ngOnInit() {
+    this.inputStreamSub = this.inputStream.pipe(debounceTime(500)).subscribe((val: string) => {
+      if (val) {
+        const inputValue = coerceNumberProperty(val, 0);
+        this.updateValueViaInput(inputValue);
+      }
+    });
+  }
+
+  public ngOnDestroy(): void {
+    this.inputStreamSub.unsubscribe();
+  }
+
+  public upBtnClicked(): void {
+    let newValue = this.value + this.step;
+    if (newValue > this.max) {
+      newValue = this.min;
     }
+    this.updateValue(newValue);
+  }
 
-    get owlDTTimerBoxClass(): boolean {
-        return true;
+  public downBtnClicked(): void {
+    let newValue = this.value - this.step;
+    if (newValue < this.min) {
+      newValue = this.max;
     }
+    this.updateValue(newValue);
+  }
 
-    public ngOnInit() {
-        this.inputStreamSub = this.inputStream
-            .pipe(debounceTime(500), distinctUntilChanged())
-            .subscribe((val: string) => {
-                if (val) {
-                    const inputValue = coerceNumberProperty(val, 0);
-                    this.updateValueViaInput(inputValue);
-                }
-            });
-    }
+  public handleInputChange(value: string): void {
+    this.inputStream.next(value);
+  }
 
-    public ngOnDestroy(): void {
-        this.inputStreamSub.unsubscribe();
+  public focusOut(value: string): void {
+    if (value) {
+      const inputValue = coerceNumberProperty(value, 0);
+      this.updateValueViaInput(inputValue);
     }
+  }
 
-    public upBtnClicked(): void {
-        this.updateValue(this.value + this.step);
+  public handleWheelChange(event: WheelEvent) {
+    const deltaY = event.deltaY;
+    if (deltaY > 0 && !this.upBtnDisabled) {
+      this.downBtnClicked();
+    } else if (deltaY < 0 && !this.downBtnDisabled) {
+      this.upBtnClicked();
     }
+  }
 
-    public downBtnClicked(): void {
-        this.updateValue(this.value - this.step);
-    }
+  private updateValue(value: number): void {
+    this.valueChange.emit(value);
+  }
 
-    public handleInputChange(value: string): void {
-        this.inputStream.next(value);
+  private updateValueViaInput(value: number): void {
+    if (value > this.max || value < this.min) {
+      return;
     }
-
-    public handleWheelChange(event: WheelEvent) {
-        const deltaY = event.deltaY;
-        if (deltaY > 0 && !this.upBtnDisabled) {
-            this.upBtnClicked();
-        } else if (deltaY < 0 && !this.downBtnDisabled) {
-            this.downBtnClicked();
-        }
-    }
-
-    private updateValue(value: number): void {
-        this.valueChange.emit(value);
-    }
-
-    private updateValueViaInput(value: number): void {
-        if (value > this.max || value < this.min) {
-            return;
-        }
-        this.inputChange.emit(value);
-    }
+    this.inputChange.emit(value);
+  }
 }
