@@ -1,15 +1,11 @@
-/**
- * date-time.module
- */
-
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { A11yModule } from '@angular/cdk/a11y';
 import { OverlayModule } from '@angular/cdk/overlay';
 import { OwlDateTimeTriggerDirective } from './date-time-picker-trigger.directive';
 import {
-    OWL_DTPICKER_SCROLL_STRATEGY_PROVIDER,
-    OwlDateTimeComponent
+  OWL_DTPICKER_SCROLL_STRATEGY_PROVIDER,
+  OwlDateTimeComponent
 } from './date-time-picker.component';
 import { OwlDateTimeContainerComponent } from './date-time-picker-container.component';
 import { OwlDateTimeInputDirective } from './date-time-picker-input.directive';
@@ -26,34 +22,33 @@ import { OwlDateTimeInlineComponent } from './date-time-inline.component';
 import { OwlDialogModule } from '../dialog/dialog.module';
 
 @NgModule({
-    imports: [CommonModule, OverlayModule, OwlDialogModule, A11yModule],
-    exports: [
-        OwlCalendarComponent,
-        OwlTimerComponent,
-        OwlDateTimeTriggerDirective,
-        OwlDateTimeInputDirective,
-        OwlDateTimeComponent,
-        OwlDateTimeInlineComponent,
-        OwlMultiYearViewComponent,
-        OwlYearViewComponent,
-        OwlMonthViewComponent
-    ],
-    declarations: [
-        OwlDateTimeTriggerDirective,
-        OwlDateTimeInputDirective,
-        OwlDateTimeComponent,
-        OwlDateTimeContainerComponent,
-        OwlMultiYearViewComponent,
-        OwlYearViewComponent,
-        OwlMonthViewComponent,
-        OwlTimerComponent,
-        OwlTimerBoxComponent,
-        OwlCalendarComponent,
-        OwlCalendarBodyComponent,
-        NumberFixedLenPipe,
-        OwlDateTimeInlineComponent
-    ],
-    providers: [OwlDateTimeIntl, OWL_DTPICKER_SCROLL_STRATEGY_PROVIDER],
-    entryComponents: [OwlDateTimeContainerComponent]
+  imports: [CommonModule, OverlayModule, OwlDialogModule, A11yModule],
+  exports: [
+    OwlCalendarComponent,
+    OwlTimerComponent,
+    OwlDateTimeTriggerDirective,
+    OwlDateTimeInputDirective,
+    OwlDateTimeComponent,
+    OwlDateTimeInlineComponent,
+    OwlMultiYearViewComponent,
+    OwlYearViewComponent,
+    OwlMonthViewComponent
+  ],
+  declarations: [
+    OwlDateTimeTriggerDirective,
+    OwlDateTimeInputDirective,
+    OwlDateTimeComponent,
+    OwlDateTimeContainerComponent,
+    OwlMultiYearViewComponent,
+    OwlYearViewComponent,
+    OwlMonthViewComponent,
+    OwlTimerComponent,
+    OwlTimerBoxComponent,
+    OwlCalendarComponent,
+    OwlCalendarBodyComponent,
+    NumberFixedLenPipe,
+    OwlDateTimeInlineComponent
+  ],
+  providers: [OwlDateTimeIntl, OWL_DTPICKER_SCROLL_STRATEGY_PROVIDER]
 })
 export class OwlDateTimeModule {}
