@@ -1,8 +1,4 @@
 /**
- * object.utils
- */
-
-/**
  * Extends an object with the *enumerable* and *own* properties of one or more source objects,
  * similar to Object.assign.
  *
@@ -10,19 +6,19 @@
  * @param sources The source objects from which properties will be copied.
  */
 export function extendObject(dest: any, ...sources: any[]): any {
-    if (dest == null) {
-        throw TypeError('Cannot convert undefined or null to object');
-    }
+  if (dest == null) {
+    throw TypeError('Cannot convert undefined or null to object');
+  }
 
-    for (const source of sources) {
-        if (source != null) {
-            for (const key in source) {
-                if (source.hasOwnProperty(key)) {
-                    dest[key] = source[key];
-                }
-            }
+  for (const source of sources) {
+    if (source != null) {
+      for (const key in source) {
+        if (source.hasOwnProperty(key)) {
+          dest[key] = source[key];
         }
+      }
     }
+  }
 
-    return dest;
+  return dest;
 }
