@@ -1,19 +1,14 @@
-/**
- * picker
- */
-
 export { OwlDateTimeModule } from './lib/date-time/date-time.module';
 export { OwlDateTimeIntl } from './lib/date-time/date-time-picker-intl.service';
-export { OwlNativeDateTimeModule } from './lib/date-time/adapter/native/native-date-time.module';
-export { OwlMomentDateTimeModule } from './lib/date-time/adapter/moment/moment-date-time.module';
-export {
-    OWL_DATE_TIME_LOCALE_PROVIDER,
-    OWL_DATE_TIME_LOCALE,
-    DateTimeAdapter
-} from './lib/date-time/adapter/date-time-adapter.class';
-export {
-    OWL_DATE_TIME_FORMATS,
-    OwlDateTimeFormats
-} from './lib/date-time/adapter/date-time-format.class';
+export { OwlNativeDateTimeModule } from './lib/adapter/native/native-date-time.module';
+export { OWL_DATE_TIME_LOCALE, DateTimeAdapter } from './lib/adapter/date-time-adapter.class';
+export { OWL_DATE_TIME_FORMATS, OwlDateTimeFormats } from './lib/adapter/date-time-format.class';
 export { OwlDateTimeInlineComponent } from './lib/date-time/date-time-inline.component';
 export { OwlDateTimeComponent } from './lib/date-time/date-time-picker.component';
+export { OwlCalendarComponent } from './lib/date-time/calendar.component';
+export { OwlTimerComponent } from './lib/date-time/timer.component';
+export { OwlDateTimeTriggerDirective } from './lib/date-time/date-time-picker-trigger.directive';
+export { OwlDateTimeInputDirective } from './lib/date-time/date-time-picker-input.directive';
+export { OwlMultiYearViewComponent } from './lib/date-time/calendar-multi-year-view.component';
+export { OwlYearViewComponent } from './lib/date-time/calendar-year-view.component';
+export { OwlMonthViewComponent } from './lib/date-time/calendar-month-view.component';
