@@ -1,31 +1,17 @@
-import { AfterViewInit, Component, ViewChild } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 
-import { Moment } from 'moment';
-import * as moment from 'moment-timezone';
-
-import { OwlDateTimeComponent } from '../../projects/picker/src/public_api';
+import { OwlDateTimeComponent } from 'picker';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.sass']
+  styleUrls: ['./app.component.scss']
 })
-export class AppComponent implements AfterViewInit {
+export class AppComponent {
   @ViewChild('date_range_component', { static: true })
-  date_range_component: OwlDateTimeComponent<AppComponent>;
-  public selectedMoments: Moment[] = [
-    moment('2019-03-11T08:00:00+11:00').tz('Australia/Sydney'),
-    moment('2019-03-11T15:00:00+11:00').tz('Australia/Sydney')
+  dateTimeComponent: OwlDateTimeComponent<AppComponent>;
+  public selectedDate: Date[] = [
+    new Date('2019-03-11T08:00:00+11:00'),
+    new Date('2019-03-11T15:00:00+11:00')
   ];
-
-  open_once = false;
-
-  ngAfterViewInit() {
-    /*
-    if (!this.open_once) {
-        this.date_range_component.open();
-        this.open_once = true;
-    }
-    */
-  }
 }
