@@ -1,22 +1,26 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NgModule } from '@angular/core';
+import { LOCALE_ID, NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 
-import { OwlDateTimeModule, OwlNativeDateTimeModule } from '../../projects/picker/src/public_api';
-
 import { AppComponent } from './app.component';
+import { OwlDateTimeModule, OwlNativeDateTimeModule } from 'picker';
 
+import { registerLocaleData } from '@angular/common';
+import localePl from '@angular/common/locales/pl';
+
+registerLocaleData(localePl);
 
 @NgModule({
-  declarations: [
-    AppComponent
-  ],
+  declarations: [AppComponent],
   imports: [
-    BrowserModule, BrowserAnimationsModule, FormsModule,
-    OwlDateTimeModule, OwlNativeDateTimeModule
+    BrowserModule,
+    BrowserAnimationsModule,
+    FormsModule,
+    OwlDateTimeModule,
+    OwlNativeDateTimeModule
   ],
-  providers: [],
+  providers: [{ provide: LOCALE_ID, useValue: 'pl' }],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
